+++ position_targets.go
@@ -95,7 +95,7 @@ func (p *Position) xrayAttacksFor(square int, piece Piece) (bitmask Bitmask) {
 	return p.attacksFor(square, piece)
 }
 
-func (p *Position) allAttacks(color uint8) (bitmask Bitmask) {
+func (p *Position) allAttacks(color int) (bitmask Bitmask) {
 	bitmask = p.pawnAttacks(color) | p.knightAttacks(color) | p.kingAttacks(color)
 
 	for bm := p.outposts[bishop(color)] | p.outposts[queen(color)]; bm.any(); bm = bm.pop() {
@@ -115,7 +115,7 @@ func (p *Position) allAttacks(color uint8) (bitmask Bitmask) {
 // This method is used in static exchange evaluation so instead of using current
 // board bitmask (p.board) we pass the one that gets continuously updated during
 // the evaluation.
-func (p *Position) attackers(color uint8, square int, board Bitmask) (bitmask Bitmask) {
+func (p *Position) attackers(color int, square int, board Bitmask) (bitmask Bitmask) {
 	bitmask  = knightMoves[square] & p.outposts[knight(color)]
 	bitmask |= maskPawn[color][square] & p.outposts[pawn(color)]
 	bitmask |= kingMoves[square] & p.outposts[king(color)]
@@ -125,7 +125,7 @@ func (p *Position) attackers(color uint8, square int, board Bitmask) (bitmask Bi
 	return bitmask
 }
 
-func (p *Position) isAttacked(color uint8, square int) bool {
+func (p *Position) isAttacked(color int, square int) bool {
 	return (knightMoves[square] & p.outposts[knight(color)]).any() ||
 	       (maskPawn[color][square] & p.outposts[pawn(color)]).any() ||
 	       (kingMoves[square] & p.outposts[king(color)]).any() ||
@@ -133,7 +133,7 @@ func (p *Position) isAttacked(color uint8, square int) bool {
 	       (p.bishopMoves(square) & (p.outposts[bishop(color)] | p.outposts[queen(color)])).any()
 }
 
-func (p *Position) pawnTargets(color uint8, pawns Bitmask) Bitmask {
+func (p *Position) pawnTargets(color int, pawns Bitmask) Bitmask {
 	if color == White {
 		return ((pawns & ^maskFile[0]) << 7) | ((pawns & ^maskFile[7]) << 9)
 	}
@@ -141,11 +141,11 @@ func (p *Position) pawnTargets(color uint8, pawns Bitmask) Bitmask {
 	return ((pawns & ^maskFile[0]) >> 9) | ((pawns & ^maskFile[7]) >> 7)
 }
 
-func (p *Position) pawnAttacks(color uint8) Bitmask {
+func (p *Position) pawnAttacks(color int) Bitmask {
 	return p.pawnTargets(color, p.outposts[pawn(color)])
 }
 
-func (p *Position) knightAttacks(color uint8) (bitmask Bitmask) {
+func (p *Position) knightAttacks(color int) (bitmask Bitmask) {
 	for bm := p.outposts[knight(color)]; bm.any(); bm = bm.pop() {
 		bitmask |= knightMoves[bm.first()]
 	}
@@ -153,7 +153,7 @@ func (p *Position) knightAttacks(color uint8) (bitmask Bitmask) {
 	return bitmask
 }
 
-func (p *Position) bishopAttacks(color uint8) (bitmask Bitmask) {
+func (p *Position) bishopAttacks(color int) (bitmask Bitmask) {
 	for bm := p.outposts[bishop(color)]; bm.any(); bm = bm.pop() {
 		bitmask |= p.bishopMoves(bm.first())
 	}
@@ -161,7 +161,7 @@ func (p *Position) bishopAttacks(color uint8) (bitmask Bitmask) {
 	return bitmask
 }
 
-func (p *Position) rookAttacks(color uint8) (bitmask Bitmask) {
+func (p *Position) rookAttacks(color int) (bitmask Bitmask) {
 	for bm := p.outposts[rook(color)]; bm.any(); bm = bm.pop() {
 		bitmask |= p.rookMoves(bm.first())
 	}
@@ -169,7 +169,7 @@ func (p *Position) rookAttacks(color uint8) (bitmask Bitmask) {
 	return bitmask
 }
 
-func (p *Position) queenAttacks(color uint8) (bitmask Bitmask) {
+func (p *Position) queenAttacks(color int) (bitmask Bitmask) {
 	for bm := p.outposts[queen(color)]; bm.any(); bm = bm.pop() {
 		square := bm.first()
 		bitmask |= p.rookMoves(square) | p.bishopMoves(square)
@@ -178,18 +178,18 @@ func (p *Position) queenAttacks(color uint8) (bitmask Bitmask) {
 	return bitmask
 }
 
-func (p *Position) kingAttacks(color uint8) Bitmask {
+func (p *Position) kingAttacks(color int) Bitmask {
 	return kingMoves[p.king[color]]
 }
 
-func (p *Position) knightAttacksAt(square int, color uint8) (bitmask Bitmask) {
+func (p *Position) knightAttacksAt(square int, color int) (bitmask Bitmask) {
 	return knightMoves[square] & ^p.outposts[color]
 }
 
-func (p *Position) bishopAttacksAt(square int, color uint8) (bitmask Bitmask) {
+func (p *Position) bishopAttacksAt(square int, color int) (bitmask Bitmask) {
 	return p.bishopMovesAt(square, p.board) & ^p.outposts[color]
 }
 
-func (p *Position) rookAttacksAt(square int, color uint8) (bitmask Bitmask) {
+func (p *Position) rookAttacksAt(square int, color int) (bitmask Bitmask) {
 	return p.rookMovesAt(square, p.board) & ^p.outposts[color]
 }
